+++ src/graphics/dxrenderer.h
@@ -49,7 +49,7 @@ public:
 protected:
     virtual bool onStartRenderer();
     virtual void onStopRenderer();
-    virtual void onRenderFrame( float currentTime, float deltaTime );
+    virtual void onRenderFrame( double currentTime, double deltaTime );
     virtual bool resizeRenderWindow( unsigned int width, unsigned int height );
 
 private:
